test: split estop clear endpoint coverage
