+++ src/p1am_control_system/backend/tests/test_backend.py
@@ -228,74 +228,6 @@ async def test_estop_trigger() -> None:
         mock_estop.assert_called_once()
 
 
-@pytest.mark.asyncio
-async def test_estop_clear_commands_plc_when_connected() -> None:
-    """POST /api/estop/clear MUST command the connected PLC to reset (#3314)."""
-    import main as backend_main
-
-    mock_clear = AsyncMock(return_value=True)
-    mock_backup_clear = AsyncMock(return_value=True)
-
-    with (
-        patch.object(modbus_manager, "_connected", True),
-        patch.object(modbus_manager, "clear_estop", mock_clear),
-        patch.object(backend_main.backup_simulator, "clear_estop", mock_backup_clear),
-    ):
-        backend_main.e_stop_active = True
-        response = client.post("/api/estop/clear")
-
-        assert response.status_code == 200
-        assert "cleared" in response.json()["message"].lower()
-        # The controller reset MUST have been commanded, not just a local flag.
-        mock_clear.assert_called_once()
-        assert backend_main.e_stop_active is False
-
-
-@pytest.mark.asyncio
-async def test_estop_clear_keeps_latch_when_plc_rejects() -> None:
-    """If the PLC does not acknowledge the reset, the HMI must stay tripped.
-
-    This is the core of #3314: the header must NOT show green 'E-STOP CLEAR'
-    while the plant remains tripped. A failed PLC reset returns an error and
-    leaves ``e_stop_active`` latched.
-    """
-    import main as backend_main
-
-    mock_clear = AsyncMock(return_value=False)
-
-    with (
-        patch.object(modbus_manager, "_connected", True),
-        patch.object(modbus_manager, "clear_estop", mock_clear),
-    ):
-        backend_main.e_stop_active = True
-        response = client.post("/api/estop/clear")
-
-        assert response.status_code == 502
-        mock_clear.assert_called_once()
-        # Latch preserved: HMI keeps showing the tripped state.
-        assert backend_main.e_stop_active is True
-
-
-@pytest.mark.asyncio
-async def test_estop_clear_uses_backup_when_plc_offline() -> None:
-    """With the PLC offline the backup simulator handles (and confirms) the reset."""
-    import main as backend_main
-
-    mock_backup_clear = AsyncMock(return_value=True)
-
-    with (
-        patch.object(modbus_manager, "_connected", False),
-        patch.object(backend_main.backup_simulator, "clear_estop", mock_backup_clear),
-    ):
-        backend_main.e_stop_active = True
-        response = client.post("/api/estop/clear")
-
-        assert response.status_code == 200
-        assert "Simulated" in response.json()["message"]
-        mock_backup_clear.assert_called_once()
-        assert backend_main.e_stop_active is False
-
-
 def test_export_data() -> None:
     """Verify historical tag states log queries.
